refactored message in package
