--- src/main/java/lo23/battleship/online/Message.java
@@ -1,13 +1,0 @@
-package lo23.battleship.online;
-
-import java.io.Serializable;
-
-/**
- * Created by xzirva on 31/10/17.
- */
-public abstract class Message implements Serializable {
-
-    String type;
-    public abstract Message process();
-}
-
