+++ src/main/java/lo23/battleship/online/network/messages/Message.java
@@ -1,8 +1,14 @@
 package lo23.battleship.online.network.messages;
 
+import java.io.Serializable;
+
 /**
- * Created by xzirva on 17/10/17.
- * TODO: To be completed
+ * Created by xzirva on 31/10/17.
  */
-public abstract class Message {
+public abstract class Message implements Serializable {
+
+    String type;
+    public abstract String getType();
+    public abstract Message process();
 }
+
